Problem: Uses old version of Google SDK
This step uses version 120 but the current
version is 151. The old version contains
among other things outdated versions of
kubectl.

Solution: Update to version 151.

Updating to 151 causes the tarball to
grow past 50MB which means wercker refuses
to deploy it as a step as the maximum
size allowed is 50MB.

To reduce the size of the step the following
was done:

* Duplicate kubectl removed, and symlink to
the gcloud one used instead;
* gsutil and bq removed from platform/ as they
are not required;
* UPX used to compress the kubectl binary